requirements.txt
python-telegram-bot
ydb
boto3